Move internal code to impl
